Remove redundant branch instructions in con_loop
Since we know the number of iterations, we can remove the redundant
three conditional branches instrucions by the technique of strip mining.

Let n be number of iterations, and k be the times of loop unrolling.
First execute n%k iterations in original loop, and then execute
remaining iterations in k unrollling loop. Process can determine whether
to take the branch or not on the head of loop body, and obviate the
checking for the rest of body, since the the remaining iterations must
be multiple of k.