Remove redundant branch instructions in loop2
Since we don't know the exact number of nodes at the beginning, we can't
use strip mining technique. However, by counting the unrolling
iterations in loop1, we can remove some branch instructions in body of
loop2.

The effect of this optimization is not very well, because we need to
add extra instrucions in loop1 and loop2 to count. Theoretically, the
effect would be more obvious with much longer list or much larger number
of unrolling.